requirements.txt
django-sass-processor
django-compressor
django-tailwind
libsass
lxml
nltk
markdown
markdown-link-attr-modifier==0.2.0
markdownify
beautifulsoup4==4.10.0
python-magic
pydub
SpeechRecognition
django-polymorphism
